Extract sample generation into function
This patch extracts the code for generating the sample output
into a function. The primary motivation is to be able to import
the function in train.py, generate the sample when evaluating
the model, and print it to the console.

Also, the code is wrapped in `try ... except KeyboardInterrupt`
to prevent backtrace overwhelming the console when the user
aborts the sample generation.