Add streaming output for sample.py
This patch adds streaming of generated tokens in `sample.py`,
to provide a better user experience - especially on a weak
machine, it could take tens of seconds to generate the full output.

It builds on https://github.com/karpathy/nanoGPT/pull/368, and refactors
the `model.generate()` method and the `generate_sample()` function to
return a generator instead of a string.

It is a bit more cumbersome to consume, but I think the value of
seeing the model generate output token by token outweights it.

The behaviour can be disabled by the `--streaming=False` parameter.