Add modular arithmetic examples generator
Adding generator of examples for modular arithmetic.

This will be useful for demonstration of synthesis and decomposition
penalties in tokenization schemas.

In the process also refactored the prepare script to taken in separate
files for tokenization and input data, this will be useful later for
testing progressive learning and curriculum training.