Add activation variants: ReLU and SquaredReLU
Adding some variations on the activation layer within the FFN (under
MLP in this codebase).

In addition to adding regular ReLU, also ading SquaredReLU.

Squared ReLU was utilized in this paper:
https://arxiv.org/pdf/2109.08668v2.pdf

And also in the Persimmon and [Fuyu](https://www.adept.ai/blog/fuyu-8b) achitectures from Adept.

It has the advantage of introducing potential sparsity via squaring the
output, which we may be able to leverage in hardware.