Organize argparse inputs for train.py
Before, to add a new flag into model variation, one had to
1. ensure GPTConfig has an option and type is set (use_feature: bool = True)
2. add feature to the argparse in train.py
3. add feature to dictionary manually in the train.py model_config

Now these steps are handled by organization of argparse.

If it is a model configuration, add to the model_group.
One still need to make sure that the option exists in the GPTConfig, but
that should be fine (one place to expose the variable, one way to
modify).

(Before, one wouldn't get any message if one hadn't manually added it to
the model_config dictionary, which also felt very manual.)

As a bonus, this also allows us an easy fix to print each of the model
settings, training settings, and logging settings before training! : D