Refactor Readme to incorporate latest feature set
This README contains a quickstart for installation, data prep, training,
smapling, and the flow for running explorations, on GPU.

Currently we have maintained compatibility as much as possible, but just
to be safe, removing the sections we have not yet tested after
refactoring the train.py and sample.py scripts to utilize argparse for
parameterization.

Also fixed the hyperlinks for both this file and the
Contributing_Features.md file.

There is a TODO added for a colab, which should be filled in once the
full url is available (maybe with the open in colabs button).