gitignore further common output directories
Probably most controversial here is the addition of `data/*/samples/*`.
I was using this to save sample prompts for datasets. Happy to drop it
if its inclusion is not desired.

The other things are all common things you'd want to gitignore: venv
directories, vs-code workspaces, output directories (using the directory
names suggested by this codebase), and the default wandb output
directory.