Add means for beta, gamma, in and out logging
Researching softmax requires that we log the inputs outputs etc.

What's left is to do the same thing for the manual softmax
implementation (with base-e) and look at only the inputs, outputs,
max-values, and denominators (inputs are interesting, outputs should
always sum to 1, max-values wonder how these compare to beta, and
denominator for comparison to output sums overtime, if shape of output
sum for constantmax is related to something with artifact of training
for example).