Fix control flow for attention mask
This change fixes a control flow bug found by our CI for window
attention and GQA.

Here we decide to make GQA not activate flash attention unless kv groups
equals number of heads.

We also decide not to initialize the window if it won't be used, which
fixes another bug.