Add Rope length to Standard Rope
Modifications allow for testing standard rope with only rotating a
subset of the rotary position embeddings.

This greatly reduces the number of calculations as well as angles
required to calculate (or pre-calculate) for standard rope embeddings.

Also added is a method to test fine-tuning of these savings into
existing networks into the module.

Saving the actual implementation in train.py for a later pull request,
as we'll also want to make this general enough for weightless updates
(like softmax and activations), as well as for swapping weigted
operations (e.g. ReLU with Layernorm or WTE Tables and Upscaling
matrices).