Update rope_sweep with random seed
Values before were strangely ordered.

It might be the case that random seeds have a lot of influence on the
final validation loss.

Adding these in in case this makes the graph more linear on average.

Also adding in a more frequent (factor of 5) increase in the validation
loss testing, in case we're just not polling the validation loss
frequent enough on this dataset.