lib: add helpers wrappers for torch.amp.autocast() and contextlib nullcontext
Some embedded GPUs like the NVIDIA Jetson nano are end of life already,
and its version of pytorch is stuck on really old versions, 1.10.0,
so there is no torch.amp.autocast() and contextlib nullcontext.

Add a simple library helpers which mask out the ugliness from the main code,
so to enable folks with these devices to keep tinkering with them.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>