lib: add helpers for tiktoken
The last library helper we need to provide a wrapper for to enable
NVIDIA Jetson Nano devices stuck on Pytorch 1.10 is for tiktoken.
We can *try* to compile it but it just fail, even on older releases,
and so let's just provide a helper which leverages the original
gpt2 encoder.py and its respective encoder.json and vocab.bpe.

To run on a Jetson nano, you can use the same exact instructions
as used for the mackbook on the README.md, just use --device=cuda

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>